requirements.txt
langchain
langchain-openai
langchain-groq
langchain-huggingface
langchain-community
langchain-text-splitters
sentence-transformers
pypdf
faiss-cpu
python-dotenv